Moved style files around
