--- pages/maps/ingestion/@id/components/cells/editor-popover.ts
@@ -1,9 +1,0 @@
-// @ts-ignore
-import hyper from "@macrostrat/hyper";
-
-import "@blueprintjs/select/lib/css/blueprint-select.css";
-import styles from "../../edit-table.module.sass";
-
-const h = hyper.styled(styles);
-
-export const EditorPopover = () => {};
