+++ pages/maps/ingestion/@id/hyper.ts
@@ -1,0 +1,10 @@
+import hyper from "@macrostrat/hyper";
+
+// Duplicative style imports
+import "~/styles/blueprint-select";
+import "@blueprintjs/select/lib/css/blueprint-select.css";
+import styles from "./tables/edit-table.module.sass";
+
+const h = hyper.styled(styles);
+
+export default h;
