+++ pages/maps/ingestion/@id/tables/edit-table.module.sass
