+++ pages/maps/ingestion/@id/tables/points.ts
@@ -13,8 +13,8 @@ import {
   CustomTableProps,
 } from "#/maps/ingestion/@id/table";
 import CheckboxCell from "#/maps/ingestion/@id/components/cells/checkbox-cell";
-import { TableInterface } from "../edit-table";
-import styles from "#/maps/ingestion/@id/edit-table.module.sass";
+import { TableInterface } from "./edit-table";
+import styles from "./edit-table.module.sass";
 import { COMMON_COLUMNS } from ".";
 import { toBoolean } from "#/maps/ingestion/@id/components/cells/util";
 import { createTableUpdate } from "#/maps/ingestion/@id/utils";
