+++ pages/maps/ingestion/@id/tables/table.d.ts
